pyproject.toml
[project]
name = "roundup"
version = "0.0.1"
description = "Default template for PDM package"
authors = [
    {name = "Giancarlo", email = "gian.verdum@gmail.com"},
]
dependencies = [
    "fastapi>=0.115.2",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.2",
    "sqlalchemy>=2.0.36",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.10",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.mypy]
packages = ["src"]
plugins = [
    "pydantic.mypy"
]
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

# isort configuration
[tool.isort]
profile = "black"  # Use black's configuration for consistent formatting
line_length = 88    # Set line length to match black's default
known_third_party = ["fastapi", "sqlalchemy"]  # Specify third-party packages
default_section = "THIRDPARTY"  # Grouping for imports
skip = []  # Add any directories/files you want to skip

[dependency-groups]
dev = ["pytest>=8.3.3", "pre-commit>=4.0.1", "mypy>=1.13.0", "isort>=5.10.0", "cookiecutter>=2.6.0", "pytest-cov>=6.0.0"]
